add delete exception

+++ src/main/java/ru/yumeno/nir/service/impl/StreetServiceImpl.java
@@ -2,8 +2,11 @@ package ru.yumeno.nir.service.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.yumeno.nir.entity.Address;
 import ru.yumeno.nir.entity.Street;
+import ru.yumeno.nir.exception_handler.exceptions.DeletionFailedException;
 import ru.yumeno.nir.exception_handler.exceptions.ResourceNotFoundException;
+import ru.yumeno.nir.repository.AddressRepository;
 import ru.yumeno.nir.repository.StreetRepository;
 import ru.yumeno.nir.service.StreetService;
 
@@ -13,10 +16,12 @@ import java.util.Optional;
 @Service
 public class StreetServiceImpl implements StreetService {
     private final StreetRepository streetRepository;
+    private final AddressRepository addressRepository;
 
     @Autowired
-    public StreetServiceImpl(StreetRepository streetRepository) {
+    public StreetServiceImpl(StreetRepository streetRepository, AddressRepository addressRepository) {
         this.streetRepository = streetRepository;
+        this.addressRepository = addressRepository;
     }
 
     @Override
@@ -43,6 +48,13 @@ public class StreetServiceImpl implements StreetService {
 
     @Override
     public void deleteStreet(int id) {
-        streetRepository.deleteById(id);
+        Optional<Street> optional = streetRepository.findById(id);
+        List<Address> addresses = addressRepository.findAllByStreet(optional.
+                orElseThrow(() -> new ResourceNotFoundException("Street not exist with id : " + id)));
+        if (addresses.isEmpty()) {
+            streetRepository.deleteById(id);
+        } else {
+            throw new DeletionFailedException("Street cannot be deleted");
+        }
     }
 }
